+++ hmath/algebra/algebra.h
@@ -21,7 +21,7 @@ namespace hmath {
     double ln(double number);
     double fmod(double number, double modulus);
     
-    double logPow(double number, int exponent, std::optional<double> modulus = std::nullopt);
+    double logPow(double number, double exponent, std::optional<double> modulus = std::nullopt);
     double sqrt(double number);
 }
